Sequence prediction
Uses Tensor, Keras, Numpy, and OS to allow user to place number inputs, then takes the mean and creates a y variable by adding those means to 1, and then to the sum of that, and so on. Then the user is prompted to input dummy numbers.  Then after a series of training and error calculation (500 epochs), the system outputs a useable set of numbers, something that might fall into a prediction that will help the user play future Powerball lotteries.