README: General information about this program
This is a project for school.  I attempted to the best of my ability to add any references utilized to create this program within all of the program header notes.  It's basically a lottery prediction program, with various menus that have different ways to look at lottery numbers.  I made it public in case any other persons new to Python and Machine Learning, as I am, might be able to use it as a helpful tool to build their own programs.  I listed references within all of the File headers, as to information I pulled from other sources.  

I individualized many of the files by building my own datasets, and utilizing user-input within the program to make predictions, etc.  The dataset is very small, so I would not say the prediction files are accurate, but the code/program works at the time of this writing.  This project is only meant for school/training purposes, as I was training on many various APIs in the making of it.